Move RObject proxy types to robj-main.ts
