+++ src/tests/packages/webr.test.ts
@@ -1,5 +1,5 @@
 import { WebR } from '../../webR/webr-main';
-import { RDouble, RLogical } from '../../webR/robj';
+import { RDouble, RLogical } from '../../webR/robj-main';
 
 const webR = new WebR({
   WEBR_URL: '../dist/',
