+++ src/tests/webR/webr-main.test.ts
@@ -11,7 +11,7 @@ import {
   REnvironment,
   RInteger,
   RFunction,
-} from '../../webR/robj';
+} from '../../webR/robj-main';
 
 const webR = new WebR({
   WEBR_URL: '../dist/',
