+++ src/webR/robj.ts
@@ -1,27 +1,9 @@
 import type { Module } from './module';
-import type { RProxy } from './proxy';
-import { isWebRPayloadPtr } from './payload';
 import { WebRDataTree, WebRDataTreeAtomic } from './tree';
+import * as RMain from './robj-main';
 import * as RWorker from './robj-worker';
 
 declare let Module: Module;
-
-// RProxy<RWorker.RObject> type aliases
-export type RObject = RProxy<RWorker.RObject>;
-export type RNull = RProxy<RWorker.RNull>;
-export type RSymbol = RProxy<RWorker.RSymbol>;
-export type RPairlist = RProxy<RWorker.RPairlist>;
-export type REnvironment = RProxy<RWorker.REnvironment>;
-export type RString = RProxy<RWorker.RString>;
-export type RLogical = RProxy<RWorker.RLogical>;
-export type RInteger = RProxy<RWorker.RInteger>;
-export type RDouble = RProxy<RWorker.RDouble>;
-export type RComplex = RProxy<RWorker.RComplex>;
-export type RCharacter = RProxy<RWorker.RCharacter>;
-export type RList = RProxy<RWorker.RList>;
-export type RRaw = RProxy<RWorker.RRaw>;
-// RFunction proxies are callable
-export type RFunction = RProxy<RWorker.RFunction> & ((...args: unknown[]) => Promise<unknown>);
 
 export type RPtr = number;
 
@@ -89,7 +71,7 @@ export type NamedObject<T> = { [key: string]: T };
  */
 export type RObjData =
   | RawType
-  | RObject
+  | RMain.RObject
   | RWorker.RObject
   | WebRDataTree
   | RObjData[]
@@ -101,33 +83,6 @@ export type RObjData =
  * associated with the vector.
  */
 export type RObjAtomicData<T> = T | (T | null)[] | WebRDataTreeAtomic<T> | NamedObject<T | null>;
-
-/**
- * Test for an RObject instance
- *
- * RObject is the user facing interface to R objects.
- *
- * @param {any} value The object to test.
- * @return {boolean} True if the object is an instance of an RObject.
- */
-export function isRObject(value: any): value is RObject {
-  return (
-    value &&
-    (typeof value === 'object' || typeof value === 'function') &&
-    'payloadType' in value &&
-    isWebRPayloadPtr(value._payload)
-  );
-}
-
-/**
- * Test for an RFunction instance
- *
- * @param {any} value The object to test.
- * @return {boolean} True if the object is an instance of an RFunction.
- */
-export function isRFunction(value: any): value is RFunction {
-  return Boolean(isRObject(value) && value._payload.obj.methods?.includes('exec'));
-}
 
 /**
  * Test if an object is of type Complex
